Refactor: 페이지 로딩 컴포넌트는 Loading, 로드 성공 컴포넌트는 Success로 통일하여 수정
