+++ client/src/page/admin/group/detail/content/GroupRemove.tsx
