+++ client/src/page/admin/member/detail/Loading.tsx
@@ -4,12 +4,12 @@ import Card from '@component/card/basic/Card';
 import CardHeader from '@component/card/basic/CardHeader';
 import CardBody from '@component/card/basic/CardBody';
 
-interface SkeletonProps {
+interface LoadingProps {
   children?: React.ReactNode;
 }
-const SkeletonDefaultProps = {};
+const LoadingDefaultProps = {};
 
-function Skeleton({ children }: SkeletonProps & typeof SkeletonDefaultProps) {
+function Loading({ children }: LoadingProps & typeof LoadingDefaultProps) {
   return (
     <>
       <section className="name-section">
@@ -43,5 +43,5 @@ function Skeleton({ children }: SkeletonProps & typeof SkeletonDefaultProps) {
   );
 }
 
-Skeleton.defaultProps = SkeletonDefaultProps;
-export default Skeleton;
+Loading.defaultProps = LoadingDefaultProps;
+export default Loading;
