+++ client/src/page/admin/member/detail/Success.tsx
@@ -5,7 +5,7 @@ import Button from '@component/form/Button';
 import Card from '@component/card/basic/Card';
 import CardHeader from '@component/card/basic/CardHeader';
 import CardBody from '@component/card/basic/CardBody';
-import MemberRemove from './MemberRemove';
+import MemberRemove from './content/MemberRemove';
 
 interface SuccessProps {
   member: typeof memberAPI.getMemberDetail.resType;
