+++ client/src/page/admin/member/detail/content/MemberRemove.tsx
